refactor(modules): extract openPopup function from main module into utils dir
