+++ src/sweetalert2.js
@@ -4,62 +4,13 @@ import { formatInputOptions, error, callIfFunction } from './utils/utils.js'
 import * as dom from './utils/dom/index'
 import setParameters from './utils/setParameters.js'
 import { DismissReason } from './utils/DismissReason'
-import {fixScrollbar} from './utils/scrollbarFix'
-import {iOSfix} from './utils/iosFix'
 import {version} from '../package.json'
 import * as staticMethods from './staticMethods'
 import * as instanceMethods from './instanceMethods'
 import globalState from './globalState'
+import { openPopup } from './utils/openPopup'
 
 let currentInstance
-
-/**
- * Animations
- *
- * @param animation
- * @param onBeforeOpen
- * @param onComplete
- */
-const openPopup = (animation, onBeforeOpen, onOpen) => {
-  const container = dom.getContainer()
-  const popup = dom.getPopup()
-
-  if (onBeforeOpen !== null && typeof onBeforeOpen === 'function') {
-    onBeforeOpen(popup)
-  }
-
-  if (animation) {
-    dom.addClass(popup, swalClasses.show)
-    dom.addClass(container, swalClasses.fade)
-    dom.removeClass(popup, swalClasses.hide)
-  } else {
-    dom.removeClass(popup, swalClasses.fade)
-  }
-  dom.show(popup)
-
-  // scrolling is 'hidden' until animation is done, after that 'auto'
-  container.style.overflowY = 'hidden'
-  if (dom.animationEndEvent && !dom.hasClass(popup, swalClasses.noanimation)) {
-    popup.addEventListener(dom.animationEndEvent, function swalCloseEventFinished () {
-      popup.removeEventListener(dom.animationEndEvent, swalCloseEventFinished)
-      container.style.overflowY = 'auto'
-    })
-  } else {
-    container.style.overflowY = 'auto'
-  }
-
-  dom.addClass([document.documentElement, document.body, container], swalClasses.shown)
-  if (dom.isModal()) {
-    fixScrollbar()
-    iOSfix()
-  }
-  dom.states.previousActiveElement = document.activeElement
-  if (onOpen !== null && typeof onOpen === 'function') {
-    setTimeout(() => {
-      onOpen(popup)
-    })
-  }
-}
 
 // SweetAlert constructor
 function SweetAlert (...args) {
