+++ src/utils/openPopup.js
@@ -1,0 +1,52 @@
+import * as dom from './dom/index'
+import { swalClasses } from './classes'
+import { fixScrollbar } from './scrollbarFix'
+import { iOSfix } from './iosFix'
+
+/**
+ * Animations
+ *
+ * @param animation
+ * @param onBeforeOpen
+ * @param onComplete
+ */
+export const openPopup = (animation, onBeforeOpen, onOpen) => {
+  const container = dom.getContainer()
+  const popup = dom.getPopup()
+
+  if (onBeforeOpen !== null && typeof onBeforeOpen === 'function') {
+    onBeforeOpen(popup)
+  }
+
+  if (animation) {
+    dom.addClass(popup, swalClasses.show)
+    dom.addClass(container, swalClasses.fade)
+    dom.removeClass(popup, swalClasses.hide)
+  } else {
+    dom.removeClass(popup, swalClasses.fade)
+  }
+  dom.show(popup)
+
+  // scrolling is 'hidden' until animation is done, after that 'auto'
+  container.style.overflowY = 'hidden'
+  if (dom.animationEndEvent && !dom.hasClass(popup, swalClasses.noanimation)) {
+    popup.addEventListener(dom.animationEndEvent, function swalCloseEventFinished () {
+      popup.removeEventListener(dom.animationEndEvent, swalCloseEventFinished)
+      container.style.overflowY = 'auto'
+    })
+  } else {
+    container.style.overflowY = 'auto'
+  }
+
+  dom.addClass([document.documentElement, document.body, container], swalClasses.shown)
+  if (dom.isModal()) {
+    fixScrollbar()
+    iOSfix()
+  }
+  dom.states.previousActiveElement = document.activeElement
+  if (onOpen !== null && typeof onOpen === 'function') {
+    setTimeout(() => {
+      onOpen(popup)
+    })
+  }
+}
